fix file-io parity regressions for output opens and dskmi2 corscl

+++ packages/py-parity-checking/sidecars/spiceypy/src/py_parity_sidecar/models/file_io.py
@@ -103,7 +103,7 @@ class StepFileIoDskmi2:
     np: int
     plates: list[int]
     finscl: float
-    corscl: float
+    corscl: int
     worksz: int
     voxpsz: int
     voxlsz: int
